+++ tests/test_stats_skill.py
@@ -9,7 +9,8 @@ from hydromt.stats import skills
 def test_skills(obsda):
     simda = obsda + 5.0
     assert np.isclose(skills.bias(simda, obsda).values, 5.0)
-    assert np.isclose(skills.percentual_bias(simda, obsda).values, 0.1033)
+    assert np.isclose(skills.percentual_bias(simda, obsda).values, 10.3299)
+    assert np.isclose(skills.volumetric_error(simda, obsda).values, 0.8967)
     assert np.isclose(skills.nashsutcliffe(simda, obsda).values, 0.97015)
     assert np.isclose(skills.lognashsutcliffe(simda, obsda).values, 0.8517)
     assert np.isclose(skills.pearson_correlation(simda, obsda).values, 1.0)
@@ -21,5 +22,6 @@ def test_skills(obsda):
         skills.kge_non_parametric_flood(simda, obsda)["kge_np_flood"].values, 0.8939
     )
     assert np.isclose(skills.rsquared(simda, obsda).values, 1.0)
-    assert np.isclose(skills.mse(simda, obsda).values, 9125.0)
-    assert np.isclose(skills.rmse(simda, obsda).values, 95.5249)
+    assert np.isclose(skills.mse(simda, obsda).values, 25.0)
+    assert np.isclose(skills.rmse(simda, obsda).values, 5.0)
+    assert np.isclose(skills.rsr(simda, obsda).values, 0.1727659)